# packages/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rust SaaS RBAC API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/sessions/password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with password & email",
        "operationId": "authenticate_with_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticateWithPasswordBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Authenticated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticateWithPasswordResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or user without password"
          },
          "403": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create account",
        "operationId": "create_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "User with same e-mail already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthenticateWithPasswordBody": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AuthenticateWithPasswordResponse": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "CreateAccountBody": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "CreateAccountResponse": {
        "type": "object",
        "required": [
          "userId"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    }
  }
}